changed to f64

+++ hostess/src/game_server.rs
@@ -35,7 +35,7 @@ pub struct Context {
     /// this value can go from close zero to many thousands 
     /// and needs to be truncated or similar by the consumer to avoid
     /// unintended behavior, e.g. players jumping through walls due to high tick
-    pub dt:f32
+    pub dt:f64
 }
 
 impl Context {
